requirements.txt
Pillow==2.6.1
twython==3.2.0
